requirements.txt
sentence-transformers==5.2.0
langchain==1.2.6
fastapi==0.128.0
pypdf==6.6.0
python-dotenv==1.2.1
langchain-pinecone==0.2.13
langchain-community==0.4.1
langchain-core == 1.2.7
streamlit == 1.53.0
groq == 1.0.0
-e .